pyproject.toml
[project]
name = "langchain-models"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flax>=0.12.0",
    "google-generativeai>=0.8.5",
    "huggingface-hub>=0.35.3",
    "langchain>=0.3.27",
    "langchain-anthropic>=0.3.21",
    "langchain-core>=0.3.78",
    "langchain-google-genai>=2.0.10",
    "langchain-huggingface>=0.3.1",
    "langchain-openai>=0.3.35",
    "numpy>=2.3.3",
    "openai>=2.2.0",
    "python-dotenv>=1.1.1",
    "scikit-learn>=1.7.2",
    "tensorflow>=2.20.0",
    "torch>=2.8.0",
    "transformers>=4.57.0",
]
